pyproject.toml
[project]
name = "dinov3-attempt2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "dotenv>=0.9.9",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyside6>=6.10.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]
